plant_disease_project/requirements.txt
tensorflow>=2.12,<3.0
opencv-python>=4.8
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
scikit-learn>=1.3
albumentations>=1.3
Pillow>=10.0
scipy>=1.10
seaborn>=0.12
streamlit>=1.29
